Agents/NetworkingResearch_Agent/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=1.0.0
ddgs>=9.0.0
